+++ pybillomat/__init__.py
@@ -6,4 +6,4 @@ Python-Billomat - Billomat API Client Library
 
 from http import Connection
 from clients import Client, Clients, ClientsIterator
-from invoices import Invoice, Invoices
+from invoices import Invoice, Invoices, InvoicesIterator