+++ history.rst
@@ -6,6 +6,15 @@ by Gerold Penz 2014
 
 
 =============
+Version 0.2.0
+=============
+
+2014-10-28
+
+- Google App Engine enabled
+
+
+=============
 Version 0.1.5
 =============